refactor(schema-diff): 모델 + 로직 클래스 4개를 별도 파일로 완전 분리
대형 파일 정리 마지막(Phase J, #206~#214에 이어). schema_diff.py(1251줄,
클래스/함수 17개)를 결합도 분석한 결과, 4개 로직 클래스(SchemaExtractor/
SchemaComparator/SeverityClassifier/SyncScriptGenerator)가 서로 전혀
생성/참조하지 않는 완전 독립 유닛이었음(diff_dialog.py/diff_workers.py가
외부에서 개별 생성). 데이터 모델 13개(Enum/dataclass)도 모든 로직
클래스가 읽기만 할 뿐 서로 얽혀있지 않아, 이전 phase들과 달리 5개
파일로 온전히 분해.

- schema_diff_models.py (신규): DiffSeverity, CompareLevel, VersionContext,
  SeveritySummary, DiffType, ColumnInfo, IndexInfo, ForeignKeyInfo,
  TableSchema, ColumnDiff, IndexDiff, ForeignKeyDiff, TableDiff,
  _normalize_column_extra
- schema_extractor.py (신규): SchemaExtractor (+ logger, 이 클래스만
  logger를 쓰고 있어 함께 이동)
- schema_comparator.py (신규): SchemaComparator
- schema_severity_classifier.py (신규): SeverityClassifier
- schema_sync_script_generator.py (신규): SyncScriptGenerator
- schema_diff.py: 1251 -> 31줄 (순수 재수입 파사드)

테스트 파일(test_schema_diff.py, test_severity_classifier.py) 전부
`from src.core.schema_diff import ...` 로만 접근하고 monkeypatch도
전혀 없어, 이번 phase는 테스트 파일 변경이 0건으로 가장 안전했음.

검증: 관련 pytest 66 passed. 전체 pytest 1809 passed / 1 known-flaky
제외. py_compile 통과. 앱 스모크 성공.

이로써 clean-code 관점 대형 파일 정리 라운드(Phase C~J, #208~#이번PR)
전부 완료.
