+++ src/core/schema_diff.py
@@ -4,1248 +4,28 @@
 - 테이블/컬럼/인덱스/FK 차이 분석
 - 동기화 SQL 스크립트 생성
 
-owner 참고: 현재 UI 노출 스키마 비교/동기화 스크립트 생성은 이 모듈의
-Python SchemaComparator / SyncScriptGenerator가 담당한다. Rust facade의
+owner 참고: 현재 UI 노출 스키마 비교/동기화 스크립트 생성은 이 모듈이
+재수입하는 SchemaComparator / SyncScriptGenerator가 담당한다. Rust facade의
 schema.diff 정리/라우팅은 이후 Rust-contract WP에서 다룬다. 이 모듈이
 DB 작업 전체를 소유한다는 의미는 아니다.
 """
-import re
-from dataclasses import dataclass, field
-from typing import List, Dict, Optional, Tuple
-from enum import Enum
-
-from src.core.logger import get_logger
-
-logger = get_logger(__name__)
-
-
-class DiffSeverity(Enum):
-    """차이 심각도"""
-    CRITICAL = "critical"   # Import 실패 위험
-    WARNING = "warning"     # 성능/무결성 영향
-    INFO = "info"           # 무시 가능
-
-
-class CompareLevel(Enum):
-    """비교 수준"""
-    QUICK = "quick"         # 테이블/컬럼 존재성, 타입만
-    STANDARD = "standard"   # + 인덱스, FK, 기본값
-    STRICT = "strict"       # + charset, collation
-
-
-@dataclass
-class VersionContext:
-    """MySQL 버전 컨텍스트"""
-    source_version: Tuple[int, int, int] = (0, 0, 0)
-    target_version: Tuple[int, int, int] = (0, 0, 0)
-    source_version_str: str = ""
-    target_version_str: str = ""
-
-
-@dataclass
-class SeveritySummary:
-    """심각도 요약"""
-    critical: int = 0
-    warning: int = 0
-    info: int = 0
-
-    @property
-    def has_critical(self) -> bool:
-        return self.critical > 0
-
-
-class DiffType(Enum):
-    """차이 유형"""
-    ADDED = "added"       # 타겟에 추가 필요
-    REMOVED = "removed"   # 타겟에서 삭제 필요
-    MODIFIED = "modified"
-    RENAMED = "renamed"   # 이름만 변경 (내용 동일)
-    UNCHANGED = "unchanged"
-
-
-def _normalize_column_extra(extra: Optional[str]) -> str:
-    """MySQL INFORMATION_SCHEMA EXTRA 값을 비교/SQL 출력용으로 정규화.
-
-    MySQL 8.0은 DEFAULT CURRENT_TIMESTAMP 등 컬럼의 EXTRA에
-    DEFAULT_GENERATED를 자동으로 붙이지만 MySQL 5.7은 붙이지 않아,
-    정규화하지 않으면 버전 차이만으로 잘못된 SQL/거짓 diff가 발생한다.
-    """
-    if not extra:
-        return ""
-    cleaned = re.sub(r"\bDEFAULT_GENERATED\b", "", extra, flags=re.IGNORECASE)
-    return " ".join(cleaned.split())
-
-
-@dataclass
-class ColumnInfo:
-    """컬럼 정보"""
-    name: str
-    data_type: str
-    nullable: bool
-    default: Optional[str]
-    extra: str = ""      # AUTO_INCREMENT 등
-    key: str = ""        # PRI, UNI, MUL
-    charset: str = ""
-    collation: str = ""
-
-    def to_sql_definition(self) -> str:
-        """SQL 컬럼 정의 생성"""
-        parts = [f"`{self.name}`", self.data_type]
-
-        if self.charset and self.charset not in self.data_type:
-            parts.append(f"CHARACTER SET {self.charset}")
-
-        if not self.nullable:
-            parts.append("NOT NULL")
-        else:
-            parts.append("NULL")
-
-        if self.default is not None:
-            if self.default.upper() in ('CURRENT_TIMESTAMP', 'NULL'):
-                parts.append(f"DEFAULT {self.default}")
-            else:
-                parts.append(f"DEFAULT '{self.default}'")
-
-        extra = _normalize_column_extra(self.extra)
-        if extra:
-            parts.append(extra)
-
-        return " ".join(parts)
-
-
-@dataclass
-class IndexInfo:
-    """인덱스 정보"""
-    name: str
-    columns: List[str]
-    unique: bool = False
-    type: str = "BTREE"   # BTREE, FULLTEXT, HASH
-
-    def to_sql_definition(self, table_name: str) -> str:
-        """인덱스 생성 SQL"""
-        cols = ", ".join(f"`{c}`" for c in self.columns)
-        if self.name == "PRIMARY":
-            return f"PRIMARY KEY ({cols})"
-        elif self.unique:
-            return f"UNIQUE INDEX `{self.name}` ({cols}) USING {self.type}"
-        else:
-            return f"INDEX `{self.name}` ({cols}) USING {self.type}"
-
-
-@dataclass
-class ForeignKeyInfo:
-    """외래 키 정보"""
-    name: str
-    columns: List[str]
-    ref_table: str
-    ref_columns: List[str]
-    on_delete: str = "RESTRICT"
-    on_update: str = "RESTRICT"
-
-    def to_sql_definition(self) -> str:
-        """FK 정의 SQL"""
-        cols = ", ".join(f"`{c}`" for c in self.columns)
-        ref_cols = ", ".join(f"`{c}`" for c in self.ref_columns)
-        return (
-            f"CONSTRAINT `{self.name}` FOREIGN KEY ({cols}) "
-            f"REFERENCES `{self.ref_table}` ({ref_cols}) "
-            f"ON DELETE {self.on_delete} ON UPDATE {self.on_update}"
-        )
-
-
-@dataclass
-class TableSchema:
-    """테이블 스키마 정보"""
-    name: str
-    columns: List[ColumnInfo] = field(default_factory=list)
-    indexes: List[IndexInfo] = field(default_factory=list)
-    foreign_keys: List[ForeignKeyInfo] = field(default_factory=list)
-    engine: str = "InnoDB"
-    charset: str = "utf8mb4"
-    collation: str = "utf8mb4_general_ci"
-    row_count: int = 0
-
-    def get_column(self, name: str) -> Optional[ColumnInfo]:
-        """이름으로 컬럼 조회"""
-        for col in self.columns:
-            if col.name.lower() == name.lower():
-                return col
-        return None
-
-    def get_index(self, name: str) -> Optional[IndexInfo]:
-        """이름으로 인덱스 조회"""
-        for idx in self.indexes:
-            if idx.name.lower() == name.lower():
-                return idx
-        return None
-
-    def get_foreign_key(self, name: str) -> Optional[ForeignKeyInfo]:
-        """이름으로 FK 조회"""
-        for fk in self.foreign_keys:
-            if fk.name.lower() == name.lower():
-                return fk
-        return None
-
-
-@dataclass
-class ColumnDiff:
-    """컬럼 차이"""
-    column_name: str
-    diff_type: DiffType
-    source_info: Optional[ColumnInfo] = None
-    target_info: Optional[ColumnInfo] = None
-    differences: List[str] = field(default_factory=list)
-    severity: Optional[DiffSeverity] = None
-
-
-@dataclass
-class IndexDiff:
-    """인덱스 차이"""
-    index_name: str
-    diff_type: DiffType
-    source_info: Optional[IndexInfo] = None
-    target_info: Optional[IndexInfo] = None
-    differences: List[str] = field(default_factory=list)
-    severity: Optional[DiffSeverity] = None
-    old_name: Optional[str] = None  # RENAMED 시 타겟 측 이전 이름
-
-
-@dataclass
-class ForeignKeyDiff:
-    """FK 차이"""
-    fk_name: str
-    diff_type: DiffType
-    source_info: Optional[ForeignKeyInfo] = None
-    target_info: Optional[ForeignKeyInfo] = None
-    differences: List[str] = field(default_factory=list)
-    severity: Optional[DiffSeverity] = None
-    old_name: Optional[str] = None  # RENAMED 시 타겟 측 이전 이름
-
-
-@dataclass
-class TableDiff:
-    """테이블 차이"""
-    table_name: str
-    diff_type: DiffType
-    source_schema: Optional[TableSchema] = None
-    target_schema: Optional[TableSchema] = None
-    column_diffs: List[ColumnDiff] = field(default_factory=list)
-    index_diffs: List[IndexDiff] = field(default_factory=list)
-    fk_diffs: List[ForeignKeyDiff] = field(default_factory=list)
-    row_count_source: int = 0
-    row_count_target: int = 0
-    severity: Optional[DiffSeverity] = None
-
-    def has_differences(self) -> bool:
-        """차이가 있는지 확인"""
-        if self.diff_type in [DiffType.ADDED, DiffType.REMOVED]:
-            return True
-        return any(d.diff_type != DiffType.UNCHANGED
-                  for d in self.column_diffs + self.index_diffs + self.fk_diffs)
-
-
-class SchemaExtractor:
-    """스키마 정보 추출기"""
-
-    def __init__(self, connector):
-        """
-        Args:
-            connector: MySQLConnector 인스턴스
-        """
-        self.connector = connector
-
-    def extract_table_schema(self, schema: str, table: str) -> Optional[TableSchema]:
-        """테이블 스키마 정보 추출
-
-        Args:
-            schema: 데이터베이스 이름
-            table: 테이블 이름
-
-        Returns:
-            TableSchema 또는 None
-        """
-        try:
-            # 컬럼 정보
-            columns = self._get_columns(schema, table)
-
-            # 인덱스 정보
-            indexes = self._get_indexes(schema, table)
-
-            # 외래 키 정보
-            foreign_keys = self._get_foreign_keys(schema, table)
-
-            # 테이블 옵션
-            engine, charset, collation = self._get_table_options(schema, table)
-
-            # 행 수
-            row_count = self._get_row_count(schema, table)
-
-            return TableSchema(
-                name=table,
-                columns=columns,
-                indexes=indexes,
-                foreign_keys=foreign_keys,
-                engine=engine,
-                charset=charset,
-                collation=collation,
-                row_count=row_count
-            )
-
-        except Exception as e:
-            logger.error(f"테이블 스키마 추출 실패 ({schema}.{table}): {e}")
-            return None
-
-    def extract_all_tables(self, schema: str) -> Dict[str, TableSchema]:
-        """스키마 내 모든 테이블 정보 추출
-
-        Args:
-            schema: 데이터베이스 이름
-
-        Returns:
-            {테이블명: TableSchema} 딕셔너리
-        """
-        tables = {}
-
-        # 테이블 목록 조회
-        query = """
-            SELECT TABLE_NAME
-            FROM INFORMATION_SCHEMA.TABLES
-            WHERE TABLE_SCHEMA = %s
-              AND TABLE_TYPE = 'BASE TABLE'
-            ORDER BY TABLE_NAME
-        """
-
-        try:
-            result = self.connector.execute(query, (schema,))
-            for row in result:
-                table_name = row['TABLE_NAME']
-                table_schema = self.extract_table_schema(schema, table_name)
-                if table_schema:
-                    tables[table_name] = table_schema
-        except Exception as e:
-            logger.error(f"테이블 목록 조회 실패 ({schema}): {e}")
-
-        return tables
-
-    def _get_columns(self, schema: str, table: str) -> List[ColumnInfo]:
-        """컬럼 정보 조회"""
-        query = """
-            SELECT
-                COLUMN_NAME,
-                COLUMN_TYPE,
-                IS_NULLABLE,
-                COLUMN_DEFAULT,
-                EXTRA,
-                COLUMN_KEY,
-                CHARACTER_SET_NAME,
-                COLLATION_NAME
-            FROM INFORMATION_SCHEMA.COLUMNS
-            WHERE TABLE_SCHEMA = %s AND TABLE_NAME = %s
-            ORDER BY ORDINAL_POSITION
-        """
-
-        columns = []
-        try:
-            result = self.connector.execute(query, (schema, table))
-            for row in result:
-                col = ColumnInfo(
-                    name=row['COLUMN_NAME'],
-                    data_type=row['COLUMN_TYPE'],
-                    nullable=(row['IS_NULLABLE'] == 'YES'),
-                    default=row['COLUMN_DEFAULT'],
-                    extra=_normalize_column_extra(row['EXTRA']),
-                    key=row['COLUMN_KEY'] or '',
-                    charset=row['CHARACTER_SET_NAME'] or '',
-                    collation=row['COLLATION_NAME'] or ''
-                )
-                columns.append(col)
-        except Exception as e:
-            logger.error(f"컬럼 정보 조회 실패: {e}")
-
-        return columns
-
-    def _get_indexes(self, schema: str, table: str) -> List[IndexInfo]:
-        """인덱스 정보 조회"""
-        query = """
-            SELECT
-                INDEX_NAME,
-                COLUMN_NAME,
-                NON_UNIQUE,
-                INDEX_TYPE
-            FROM INFORMATION_SCHEMA.STATISTICS
-            WHERE TABLE_SCHEMA = %s AND TABLE_NAME = %s
-            ORDER BY INDEX_NAME, SEQ_IN_INDEX
-        """
-
-        index_map = {}
-        try:
-            result = self.connector.execute(query, (schema, table))
-            for row in result:
-                idx_name = row['INDEX_NAME']
-                col_name = row['COLUMN_NAME']
-                non_unique = row['NON_UNIQUE']
-                idx_type = row['INDEX_TYPE']
-
-                if idx_name not in index_map:
-                    index_map[idx_name] = IndexInfo(
-                        name=idx_name,
-                        columns=[],
-                        unique=(non_unique == 0),
-                        type=idx_type
-                    )
-                index_map[idx_name].columns.append(col_name)
-        except Exception as e:
-            logger.error(f"인덱스 정보 조회 실패: {e}")
-
-        return list(index_map.values())
-
-    def _get_foreign_keys(self, schema: str, table: str) -> List[ForeignKeyInfo]:
-        """외래 키 정보 조회"""
-        query = """
-            SELECT
-                kcu.CONSTRAINT_NAME,
-                kcu.COLUMN_NAME,
-                kcu.REFERENCED_TABLE_NAME,
-                kcu.REFERENCED_COLUMN_NAME,
-                rc.DELETE_RULE,
-                rc.UPDATE_RULE
-            FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE kcu
-            JOIN INFORMATION_SCHEMA.REFERENTIAL_CONSTRAINTS rc
-                ON kcu.CONSTRAINT_NAME = rc.CONSTRAINT_NAME
-                AND kcu.TABLE_SCHEMA = rc.CONSTRAINT_SCHEMA
-            WHERE kcu.TABLE_SCHEMA = %s
-              AND kcu.TABLE_NAME = %s
-              AND kcu.REFERENCED_TABLE_NAME IS NOT NULL
-            ORDER BY kcu.CONSTRAINT_NAME, kcu.ORDINAL_POSITION
-        """
-
-        fk_map = {}
-        try:
-            result = self.connector.execute(query, (schema, table))
-            for row in result:
-                fk_name = row['CONSTRAINT_NAME']
-                col = row['COLUMN_NAME']
-                ref_table = row['REFERENCED_TABLE_NAME']
-                ref_col = row['REFERENCED_COLUMN_NAME']
-                on_del = row['DELETE_RULE']
-                on_upd = row['UPDATE_RULE']
-
-                if fk_name not in fk_map:
-                    fk_map[fk_name] = ForeignKeyInfo(
-                        name=fk_name,
-                        columns=[],
-                        ref_table=ref_table,
-                        ref_columns=[],
-                        on_delete=on_del,
-                        on_update=on_upd
-                    )
-                fk_map[fk_name].columns.append(col)
-                fk_map[fk_name].ref_columns.append(ref_col)
-        except Exception as e:
-            logger.error(f"FK 정보 조회 실패: {e}")
-
-        return list(fk_map.values())
-
-    def _get_table_options(self, schema: str, table: str) -> Tuple[str, str, str]:
-        """테이블 옵션 조회"""
-        query = """
-            SELECT ENGINE, TABLE_COLLATION
-            FROM INFORMATION_SCHEMA.TABLES
-            WHERE TABLE_SCHEMA = %s AND TABLE_NAME = %s
-        """
-
-        try:
-            result = self.connector.execute(query, (schema, table))
-            if result:
-                row = result[0]
-                engine = row['ENGINE']
-                collation = row['TABLE_COLLATION']
-
-                # Collation에서 charset 추출
-                charset = collation.split('_')[0] if collation else 'utf8mb4'
-                return engine or 'InnoDB', charset, collation or ''
-        except Exception as e:
-            logger.error(f"테이블 옵션 조회 실패: {e}")
-
-        return 'InnoDB', 'utf8mb4', 'utf8mb4_general_ci'
-
-    def _get_row_count(self, schema: str, table: str) -> int:
-        """테이블 행 수 조회"""
-        query = f"SELECT COUNT(*) as cnt FROM `{schema}`.`{table}`"
-        try:
-            result = self.connector.execute(query)
-            if result:
-                return result[0]['cnt']
-        except Exception:
-            pass
-        return 0
-
-
-class SchemaComparator:
-    """스키마 비교기"""
-
-    def compare_tables(
-        self,
-        source: TableSchema,
-        target: TableSchema,
-        compare_level: CompareLevel = CompareLevel.STANDARD
-    ) -> TableDiff:
-        """두 테이블 스키마 비교
-
-        Args:
-            source: 소스 테이블 스키마
-            target: 타겟 테이블 스키마
-            compare_level: 비교 수준
-
-        Returns:
-            TableDiff
-        """
-        diff = TableDiff(
-            table_name=source.name,
-            diff_type=DiffType.UNCHANGED,
-            source_schema=source,
-            target_schema=target,
-            row_count_source=source.row_count,
-            row_count_target=target.row_count
-        )
-
-        # 컬럼 비교
-        diff.column_diffs = self._compare_columns(
-            source.columns, target.columns, compare_level
-        )
-
-        # 인덱스/FK 비교 (Quick 모드에서는 스킵)
-        if compare_level in (CompareLevel.STANDARD, CompareLevel.STRICT):
-            diff.index_diffs = self._compare_indexes(source.indexes, target.indexes)
-            diff.fk_diffs = self._compare_foreign_keys(
-                source.foreign_keys, target.foreign_keys
-            )
-
-        # 전체 상태 결정
-        if diff.has_differences():
-            diff.diff_type = DiffType.MODIFIED
-
-        return diff
-
-    def compare_schemas(
-        self,
-        source_tables: Dict[str, TableSchema],
-        target_tables: Dict[str, TableSchema],
-        compare_level: CompareLevel = CompareLevel.STANDARD
-    ) -> List[TableDiff]:
-        """두 스키마 전체 비교
-
-        Args:
-            source_tables: 소스 테이블 딕셔너리
-            target_tables: 타겟 테이블 딕셔너리
-            compare_level: 비교 수준
-
-        Returns:
-            TableDiff 목록
-        """
-        diffs = []
-
-        all_tables = set(source_tables.keys()) | set(target_tables.keys())
-
-        for table_name in sorted(all_tables):
-            source = source_tables.get(table_name)
-            target = target_tables.get(table_name)
-
-            if source and not target:
-                # 소스에만 있음 (타겟에 추가 필요)
-                diff = TableDiff(
-                    table_name=table_name,
-                    diff_type=DiffType.ADDED,
-                    source_schema=source,
-                    row_count_source=source.row_count
-                )
-            elif target and not source:
-                # 타겟에만 있음 (삭제 필요)
-                diff = TableDiff(
-                    table_name=table_name,
-                    diff_type=DiffType.REMOVED,
-                    target_schema=target,
-                    row_count_target=target.row_count
-                )
-            else:
-                # 둘 다 있음 (상세 비교)
-                diff = self.compare_tables(source, target, compare_level)
-
-            diffs.append(diff)
-
-        return diffs
-
-    def _compare_columns(
-        self,
-        source_cols: List[ColumnInfo],
-        target_cols: List[ColumnInfo],
-        compare_level: CompareLevel = CompareLevel.STANDARD
-    ) -> List[ColumnDiff]:
-        """컬럼 비교"""
-        diffs = []
-
-        source_map = {c.name.lower(): c for c in source_cols}
-        target_map = {c.name.lower(): c for c in target_cols}
-
-        all_cols = set(source_map.keys()) | set(target_map.keys())
-
-        for col_name in sorted(all_cols):
-            src = source_map.get(col_name)
-            tgt = target_map.get(col_name)
-
-            if src and not tgt:
-                diffs.append(ColumnDiff(
-                    column_name=src.name,
-                    diff_type=DiffType.ADDED,
-                    source_info=src
-                ))
-            elif tgt and not src:
-                diffs.append(ColumnDiff(
-                    column_name=tgt.name,
-                    diff_type=DiffType.REMOVED,
-                    target_info=tgt
-                ))
-            else:
-                # 상세 비교
-                differences = []
-
-                # 타입 비교 (모든 레벨)
-                if src.data_type.lower() != tgt.data_type.lower():
-                    differences.append(f"타입: {src.data_type} → {tgt.data_type}")
-
-                # Quick 모드: 타입만 비교
-                if compare_level != CompareLevel.QUICK:
-                    if src.nullable != tgt.nullable:
-                        src_null = "NULL" if src.nullable else "NOT NULL"
-                        tgt_null = "NULL" if tgt.nullable else "NOT NULL"
-                        differences.append(f"Nullable: {src_null} → {tgt_null}")
-
-                    if src.default != tgt.default:
-                        differences.append(f"Default: {src.default} → {tgt.default}")
-
-                    src_extra = _normalize_column_extra(src.extra)
-                    tgt_extra = _normalize_column_extra(tgt.extra)
-                    if src_extra.lower() != tgt_extra.lower():
-                        differences.append(f"Extra: {src_extra} → {tgt_extra}")
-
-                # Strict 모드: charset + collation 추가 비교
-                if compare_level == CompareLevel.STRICT:
-                    if src.charset and tgt.charset and src.charset.lower() != tgt.charset.lower():
-                        differences.append(f"Charset: {src.charset} → {tgt.charset}")
-
-                    if src.collation and tgt.collation and src.collation.lower() != tgt.collation.lower():
-                        differences.append(f"Collation: {src.collation} → {tgt.collation}")
-
-                if differences:
-                    diffs.append(ColumnDiff(
-                        column_name=src.name,
-                        diff_type=DiffType.MODIFIED,
-                        source_info=src,
-                        target_info=tgt,
-                        differences=differences
-                    ))
-                else:
-                    diffs.append(ColumnDiff(
-                        column_name=src.name,
-                        diff_type=DiffType.UNCHANGED,
-                        source_info=src,
-                        target_info=tgt
-                    ))
-
-        return diffs
-
-    @staticmethod
-    def _index_content_key(idx: IndexInfo) -> tuple:
-        """인덱스의 내용 기반 키 (이름 제외)"""
-        return (tuple(c.lower() for c in idx.columns), idx.unique, idx.type)
-
-    @staticmethod
-    def _fk_content_key(fk: ForeignKeyInfo) -> tuple:
-        """FK의 내용 기반 키 (이름 제외)"""
-        return (
-            tuple(c.lower() for c in fk.columns),
-            fk.ref_table.lower(),
-            tuple(c.lower() for c in fk.ref_columns),
-            fk.on_delete,
-            fk.on_update,
-        )
-
-    def _compare_indexes(
-        self,
-        source_idx: List[IndexInfo],
-        target_idx: List[IndexInfo]
-    ) -> List[IndexDiff]:
-        """인덱스 비교 (rename 감지 포함)"""
-        diffs = []
-
-        source_map = {i.name.lower(): i for i in source_idx}
-        target_map = {i.name.lower(): i for i in target_idx}
-
-        # 1단계: 이름으로 매칭
-        matched_source = set()
-        matched_target = set()
-
-        common_names = set(source_map.keys()) & set(target_map.keys())
-        for idx_name in sorted(common_names):
-            src = source_map[idx_name]
-            tgt = target_map[idx_name]
-            matched_source.add(idx_name)
-            matched_target.add(idx_name)
-
-            differences = []
-            if src.columns != tgt.columns:
-                differences.append(f"컬럼: {src.columns} → {tgt.columns}")
-            if src.unique != tgt.unique:
-                differences.append(f"Unique: {src.unique} → {tgt.unique}")
-
-            if differences:
-                diffs.append(IndexDiff(
-                    index_name=src.name,
-                    diff_type=DiffType.MODIFIED,
-                    source_info=src,
-                    target_info=tgt,
-                    differences=differences
-                ))
-            else:
-                diffs.append(IndexDiff(
-                    index_name=src.name,
-                    diff_type=DiffType.UNCHANGED,
-                    source_info=src,
-                    target_info=tgt
-                ))
-
-        # 2단계: 미매칭 항목에서 rename 감지
-        unmatched_source = {k: v for k, v in source_map.items() if k not in matched_source}
-        unmatched_target = {k: v for k, v in target_map.items() if k not in matched_target}
-
-        # 타겟 미매칭을 내용 기반으로 인덱싱
-        target_by_content = {}
-        for tgt_name, tgt in unmatched_target.items():
-            key = self._index_content_key(tgt)
-            target_by_content.setdefault(key, []).append(tgt_name)
-
-        renamed_target = set()
-        source_added = []
-
-        for src_name in sorted(unmatched_source.keys()):
-            src = unmatched_source[src_name]
-            content_key = self._index_content_key(src)
-            candidates = target_by_content.get(content_key, [])
-            # 아직 매칭 안 된 후보 찾기
-            match_found = False
-            for tgt_name in candidates:
-                if tgt_name not in renamed_target:
-                    tgt = unmatched_target[tgt_name]
-                    renamed_target.add(tgt_name)
-                    diffs.append(IndexDiff(
-                        index_name=src.name,
-                        diff_type=DiffType.RENAMED,
-                        source_info=src,
-                        target_info=tgt,
-                        differences=[f"이름 변경: {tgt.name} → {src.name}"],
-                        old_name=tgt.name
-                    ))
-                    match_found = True
-                    break
-
-            if not match_found:
-                source_added.append(src)
-
-        # 3단계: 남은 미매칭 → ADDED / REMOVED
-        for src in source_added:
-            diffs.append(IndexDiff(
-                index_name=src.name,
-                diff_type=DiffType.ADDED,
-                source_info=src
-            ))
-
-        for tgt_name in sorted(unmatched_target.keys()):
-            if tgt_name not in renamed_target:
-                tgt = unmatched_target[tgt_name]
-                diffs.append(IndexDiff(
-                    index_name=tgt.name,
-                    diff_type=DiffType.REMOVED,
-                    target_info=tgt
-                ))
-
-        return diffs
-
-    def _compare_foreign_keys(
-        self,
-        source_fks: List[ForeignKeyInfo],
-        target_fks: List[ForeignKeyInfo]
-    ) -> List[ForeignKeyDiff]:
-        """외래 키 비교 (rename 감지 포함)"""
-        diffs = []
-
-        source_map = {f.name.lower(): f for f in source_fks}
-        target_map = {f.name.lower(): f for f in target_fks}
-
-        # 1단계: 이름으로 매칭
-        matched_source = set()
-        matched_target = set()
-
-        common_names = set(source_map.keys()) & set(target_map.keys())
-        for fk_name in sorted(common_names):
-            src = source_map[fk_name]
-            tgt = target_map[fk_name]
-            matched_source.add(fk_name)
-            matched_target.add(fk_name)
-
-            differences = []
-            if src.ref_table != tgt.ref_table:
-                differences.append(f"참조 테이블: {src.ref_table} → {tgt.ref_table}")
-            if src.columns != tgt.columns:
-                differences.append(f"컬럼: {src.columns} → {tgt.columns}")
-            if src.on_delete != tgt.on_delete:
-                differences.append(f"ON DELETE: {src.on_delete} → {tgt.on_delete}")
-            if src.on_update != tgt.on_update:
-                differences.append(f"ON UPDATE: {src.on_update} → {tgt.on_update}")
-
-            if differences:
-                diffs.append(ForeignKeyDiff(
-                    fk_name=src.name,
-                    diff_type=DiffType.MODIFIED,
-                    source_info=src,
-                    target_info=tgt,
-                    differences=differences
-                ))
-            else:
-                diffs.append(ForeignKeyDiff(
-                    fk_name=src.name,
-                    diff_type=DiffType.UNCHANGED,
-                    source_info=src,
-                    target_info=tgt
-                ))
-
-        # 2단계: 미매칭 항목에서 rename 감지
-        unmatched_source = {k: v for k, v in source_map.items() if k not in matched_source}
-        unmatched_target = {k: v for k, v in target_map.items() if k not in matched_target}
-
-        target_by_content = {}
-        for tgt_name, tgt in unmatched_target.items():
-            key = self._fk_content_key(tgt)
-            target_by_content.setdefault(key, []).append(tgt_name)
-
-        renamed_target = set()
-        source_added = []
-
-        for src_name in sorted(unmatched_source.keys()):
-            src = unmatched_source[src_name]
-            content_key = self._fk_content_key(src)
-            candidates = target_by_content.get(content_key, [])
-
-            match_found = False
-            for tgt_name in candidates:
-                if tgt_name not in renamed_target:
-                    tgt = unmatched_target[tgt_name]
-                    renamed_target.add(tgt_name)
-                    diffs.append(ForeignKeyDiff(
-                        fk_name=src.name,
-                        diff_type=DiffType.RENAMED,
-                        source_info=src,
-                        target_info=tgt,
-                        differences=[f"이름 변경: {tgt.name} → {src.name}"],
-                        old_name=tgt.name
-                    ))
-                    match_found = True
-                    break
-
-            if not match_found:
-                source_added.append(src)
-
-        # 3단계: 남은 미매칭 → ADDED / REMOVED
-        for src in source_added:
-            diffs.append(ForeignKeyDiff(
-                fk_name=src.name,
-                diff_type=DiffType.ADDED,
-                source_info=src
-            ))
-
-        for tgt_name in sorted(unmatched_target.keys()):
-            if tgt_name not in renamed_target:
-                tgt = unmatched_target[tgt_name]
-                diffs.append(ForeignKeyDiff(
-                    fk_name=tgt.name,
-                    diff_type=DiffType.REMOVED,
-                    target_info=tgt
-                ))
-
-        return diffs
-
-
-class SeverityClassifier:
-    """비교 결과에 심각도를 부여하는 후처리 레이어"""
-
-    # display width만 다른 integer 계열 타입
-    _INTEGER_TYPES = frozenset([
-        'tinyint', 'smallint', 'mediumint', 'int', 'integer', 'bigint'
-    ])
-
-    def __init__(self, version_context: Optional[VersionContext] = None):
-        self.version_context = version_context
-
-    def classify(
-        self, diffs: List[TableDiff]
-    ) -> Tuple[List[TableDiff], SeveritySummary]:
-        """비교 결과에 심각도 부여
-
-        Args:
-            diffs: TableDiff 목록
-
-        Returns:
-            (심각도 부여된 diffs, SeveritySummary)
-        """
-        summary = SeveritySummary()
-
-        for table_diff in diffs:
-            # 테이블 레벨 심각도
-            table_diff.severity = self._classify_table(table_diff)
-            if table_diff.severity:
-                self._count(summary, table_diff.severity)
-
-            # 컬럼 심각도
-            for col_diff in table_diff.column_diffs:
-                col_diff.severity = self._classify_column(col_diff)
-                if col_diff.severity:
-                    self._count(summary, col_diff.severity)
-
-            # 인덱스 심각도
-            for idx_diff in table_diff.index_diffs:
-                idx_diff.severity = self._classify_index(idx_diff)
-                if idx_diff.severity:
-                    self._count(summary, idx_diff.severity)
-
-            # FK 심각도
-            for fk_diff in table_diff.fk_diffs:
-                fk_diff.severity = self._classify_fk(fk_diff)
-                if fk_diff.severity:
-                    self._count(summary, fk_diff.severity)
-
-        return diffs, summary
-
-    def _count(self, summary: SeveritySummary, severity: DiffSeverity):
-        """심각도 카운트 증가"""
-        if severity == DiffSeverity.CRITICAL:
-            summary.critical += 1
-        elif severity == DiffSeverity.WARNING:
-            summary.warning += 1
-        elif severity == DiffSeverity.INFO:
-            summary.info += 1
-
-    def _classify_table(self, diff: TableDiff) -> Optional[DiffSeverity]:
-        """테이블 레벨 심각도 분류"""
-        if diff.diff_type in (DiffType.ADDED, DiffType.REMOVED):
-            return DiffSeverity.CRITICAL
-        return None
-
-    def _classify_column(self, diff: ColumnDiff) -> Optional[DiffSeverity]:
-        """컬럼 심각도 분류"""
-        if diff.diff_type == DiffType.UNCHANGED:
-            return None
-        if diff.diff_type in (DiffType.ADDED, DiffType.REMOVED):
-            return DiffSeverity.CRITICAL
-
-        # MODIFIED: 각 변경 항목별 심각도 판단
-        max_severity = None
-
-        for d in diff.differences:
-            if d.startswith("타입:"):
-                sev = self._classify_type_change(d, diff)
-            elif d.startswith("Nullable:"):
-                sev = DiffSeverity.WARNING
-            elif d.startswith("Default:"):
-                sev = DiffSeverity.WARNING
-            elif d.startswith("Extra:"):
-                sev = self._classify_extra_change(d)
-            elif d.startswith("Charset:") or d.startswith("Collation:"):
-                sev = DiffSeverity.WARNING
-            else:
-                sev = DiffSeverity.WARNING
-
-            max_severity = self._max_severity(max_severity, sev)
-
-        return max_severity
-
-    def _classify_type_change(
-        self, diff_text: str, col_diff: ColumnDiff
-    ) -> DiffSeverity:
-        """타입 변경 심각도 판단"""
-        src_type = col_diff.source_info.data_type if col_diff.source_info else ""
-        tgt_type = col_diff.target_info.data_type if col_diff.target_info else ""
-
-        if self._is_display_width_only_diff(src_type, tgt_type):
-            return DiffSeverity.INFO
-
-        # base type이 다르면 Critical
-        src_base = re.sub(r'\(.*?\)', '', src_type).strip().lower()
-        tgt_base = re.sub(r'\(.*?\)', '', tgt_type).strip().lower()
-
-        if src_base != tgt_base:
-            return DiffSeverity.CRITICAL
-
-        # 같은 base type이지만 size/precision이 다른 경우
-        return DiffSeverity.WARNING
-
-    def _is_display_width_only_diff(self, src: str, tgt: str) -> bool:
-        """int(11) vs int 같은 display width만 다른지 확인"""
-        # display width 제거 후 비교
-        src_stripped = re.sub(r'\(\d+\)', '', src).strip().lower()
-        tgt_stripped = re.sub(r'\(\d+\)', '', tgt).strip().lower()
-
-        if src_stripped != tgt_stripped:
-            return False
-
-        # base type이 integer 계열인지 확인
-        # "int unsigned" -> "int"
-        base_type = src_stripped.split()[0] if src_stripped else ""
-        return base_type in self._INTEGER_TYPES
-
-    def _classify_extra_change(self, diff_text: str) -> DiffSeverity:
-        """Extra 필드 변경 심각도 판단"""
-        diff_lower = diff_text.lower()
-        if 'auto_increment' in diff_lower:
-            return DiffSeverity.CRITICAL
-        return DiffSeverity.WARNING
-
-    def _classify_index(self, diff: IndexDiff) -> Optional[DiffSeverity]:
-        """인덱스 심각도 분류"""
-        if diff.diff_type == DiffType.UNCHANGED:
-            return None
-
-        # 이름만 변경은 Info (내용 동일)
-        if diff.diff_type == DiffType.RENAMED:
-            return DiffSeverity.INFO
-
-        # PRIMARY KEY 변경은 Critical
-        if diff.index_name.upper() == 'PRIMARY':
-            return DiffSeverity.CRITICAL
-
-        # 기타 인덱스 변경은 Warning
-        return DiffSeverity.WARNING
-
-    def _classify_fk(self, diff: ForeignKeyDiff) -> Optional[DiffSeverity]:
-        """FK 심각도 분류"""
-        if diff.diff_type == DiffType.UNCHANGED:
-            return None
-
-        # 이름만 변경은 Info (내용 동일)
-        if diff.diff_type == DiffType.RENAMED:
-            return DiffSeverity.INFO
-
-        return DiffSeverity.WARNING
-
-    def _max_severity(
-        self,
-        current: Optional[DiffSeverity],
-        new: Optional[DiffSeverity]
-    ) -> Optional[DiffSeverity]:
-        """두 심각도 중 더 높은 값 반환"""
-        if current is None:
-            return new
-        if new is None:
-            return current
-
-        order = {
-            DiffSeverity.CRITICAL: 3,
-            DiffSeverity.WARNING: 2,
-            DiffSeverity.INFO: 1,
-        }
-        if order.get(new, 0) > order.get(current, 0):
-            return new
-        return current
-
-
-class SyncScriptGenerator:
-    """동기화 SQL 스크립트 생성기"""
-
-    def generate_sync_script(self, diffs: List[TableDiff], target_schema: str) -> str:
-        """전체 동기화 스크립트 생성
-
-        Args:
-            diffs: TableDiff 목록
-            target_schema: 타겟 스키마 이름
-
-        Returns:
-            SQL 스크립트
-        """
-        lines = [
-            "-- =======================================================",
-            "-- 스키마 동기화 스크립트",
-            f"-- 타겟: {target_schema}",
-            "-- 주의: 실행 전 반드시 백업을 수행하세요!",
-            "-- =======================================================",
-            "",
-            "SET FOREIGN_KEY_CHECKS = 0;",
-            ""
-        ]
-
-        # 1. FK 삭제 (의존성 해제)
-        fk_drops = []
-        for diff in diffs:
-            if diff.diff_type == DiffType.REMOVED:
-                if diff.target_schema:
-                    for fk in diff.target_schema.foreign_keys:
-                        fk_drops.append(
-                            f"ALTER TABLE `{target_schema}`.`{diff.table_name}` "
-                            f"DROP FOREIGN KEY `{fk.name}`;"
-                        )
-            elif diff.diff_type == DiffType.MODIFIED:
-                for fk_diff in diff.fk_diffs:
-                    if fk_diff.diff_type in [DiffType.REMOVED, DiffType.MODIFIED]:
-                        fk_drops.append(
-                            f"ALTER TABLE `{target_schema}`.`{diff.table_name}` "
-                            f"DROP FOREIGN KEY `{fk_diff.fk_name}`;"
-                        )
-                    elif fk_diff.diff_type == DiffType.RENAMED and fk_diff.old_name:
-                        # FK rename = DROP old + ADD new (MySQL에 RENAME FK 없음)
-                        fk_drops.append(
-                            f"ALTER TABLE `{target_schema}`.`{diff.table_name}` "
-                            f"DROP FOREIGN KEY `{fk_diff.old_name}`; "
-                            f"-- renamed → {fk_diff.fk_name}"
-                        )
-
-        if fk_drops:
-            lines.append("-- FK 삭제")
-            lines.extend(fk_drops)
-            lines.append("")
-
-        # 2. 테이블 삭제 (소스에 없는 테이블)
-        table_drops = []
-        for diff in diffs:
-            if diff.diff_type == DiffType.REMOVED:
-                table_drops.append(f"DROP TABLE IF EXISTS `{target_schema}`.`{diff.table_name}`;")
-
-        if table_drops:
-            lines.append("-- 테이블 삭제")
-            lines.extend(table_drops)
-            lines.append("")
-
-        # 3. 테이블 생성 (타겟에 없는 테이블)
-        table_creates = []
-        for diff in diffs:
-            if diff.diff_type == DiffType.ADDED and diff.source_schema:
-                table_creates.append(
-                    self._generate_create_table(target_schema, diff.source_schema)
-                )
-
-        if table_creates:
-            lines.append("-- 테이블 생성")
-            lines.extend(table_creates)
-            lines.append("")
-
-        # 4. 컬럼/인덱스 변경
-        alter_statements = []
-        for diff in diffs:
-            if diff.diff_type == DiffType.MODIFIED:
-                # 컬럼 변경
-                for col_diff in diff.column_diffs:
-                    if col_diff.diff_type == DiffType.ADDED and col_diff.source_info:
-                        alter_statements.append(
-                            f"ALTER TABLE `{target_schema}`.`{diff.table_name}` "
-                            f"ADD COLUMN {col_diff.source_info.to_sql_definition()};"
-                        )
-                    elif col_diff.diff_type == DiffType.REMOVED:
-                        alter_statements.append(
-                            f"ALTER TABLE `{target_schema}`.`{diff.table_name}` "
-                            f"DROP COLUMN `{col_diff.column_name}`;"
-                        )
-                    elif col_diff.diff_type == DiffType.MODIFIED and col_diff.source_info:
-                        alter_statements.append(
-                            f"ALTER TABLE `{target_schema}`.`{diff.table_name}` "
-                            f"MODIFY COLUMN {col_diff.source_info.to_sql_definition()};"
-                        )
-
-                # 인덱스 변경
-                for idx_diff in diff.index_diffs:
-                    if idx_diff.index_name == 'PRIMARY':
-                        continue  # PRIMARY KEY는 별도 처리 필요
-
-                    if idx_diff.diff_type == DiffType.ADDED and idx_diff.source_info:
-                        idx_sql = idx_diff.source_info.to_sql_definition(diff.table_name)
-                        alter_statements.append(
-                            f"ALTER TABLE `{target_schema}`.`{diff.table_name}` ADD {idx_sql};"
-                        )
-                    elif idx_diff.diff_type == DiffType.REMOVED:
-                        alter_statements.append(
-                            f"ALTER TABLE `{target_schema}`.`{diff.table_name}` "
-                            f"DROP INDEX `{idx_diff.index_name}`;"
-                        )
-                    elif idx_diff.diff_type == DiffType.MODIFIED and idx_diff.source_info:
-                        # 인덱스 수정 = 삭제 후 재생성
-                        alter_statements.append(
-                            f"ALTER TABLE `{target_schema}`.`{diff.table_name}` "
-                            f"DROP INDEX `{idx_diff.index_name}`;"
-                        )
-                        idx_sql = idx_diff.source_info.to_sql_definition(diff.table_name)
-                        alter_statements.append(
-                            f"ALTER TABLE `{target_schema}`.`{diff.table_name}` ADD {idx_sql};"
-                        )
-                    elif idx_diff.diff_type == DiffType.RENAMED and idx_diff.old_name:
-                        # MySQL 5.7+ RENAME INDEX
-                        alter_statements.append(
-                            f"ALTER TABLE `{target_schema}`.`{diff.table_name}` "
-                            f"RENAME INDEX `{idx_diff.old_name}` TO `{idx_diff.index_name}`;"
-                        )
-
-        if alter_statements:
-            lines.append("-- 컬럼/인덱스 변경")
-            lines.extend(alter_statements)
-            lines.append("")
-
-        # 5. FK 추가 (의존성 복원)
-        fk_adds = []
-        for diff in diffs:
-            if diff.diff_type == DiffType.ADDED and diff.source_schema:
-                for fk in diff.source_schema.foreign_keys:
-                    fk_adds.append(
-                        f"ALTER TABLE `{target_schema}`.`{diff.table_name}` "
-                        f"ADD {fk.to_sql_definition()};"
-                    )
-            elif diff.diff_type == DiffType.MODIFIED:
-                for fk_diff in diff.fk_diffs:
-                    if fk_diff.diff_type in [DiffType.ADDED, DiffType.MODIFIED, DiffType.RENAMED]:
-                        if fk_diff.source_info:
-                            fk_adds.append(
-                                f"ALTER TABLE `{target_schema}`.`{diff.table_name}` "
-                                f"ADD {fk_diff.source_info.to_sql_definition()};"
-                            )
-
-        if fk_adds:
-            lines.append("-- FK 추가")
-            lines.extend(fk_adds)
-            lines.append("")
-
-        lines.append("SET FOREIGN_KEY_CHECKS = 1;")
-        lines.append("")
-        lines.append("-- 스크립트 끝")
-
-        return "\n".join(lines)
-
-    def _generate_create_table(self, schema: str, table: TableSchema) -> str:
-        """CREATE TABLE 문 생성"""
-        lines = [f"CREATE TABLE `{schema}`.`{table.name}` ("]
-
-        # 컬럼
-        col_defs = [f"    {col.to_sql_definition()}" for col in table.columns]
-
-        # PRIMARY KEY
-        # 컬럼 ordinal 스캔이 아닌 PRIMARY 인덱스 메타데이터(SEQ_IN_INDEX 순서)로
-        # 생성해야 복합 PK의 실제 컬럼 순서가 보존된다.
-        primary_idx = table.get_index("PRIMARY")
-        if primary_idx and primary_idx.columns:
-            col_defs.append(f"    {primary_idx.to_sql_definition(table.name)}")
-
-        # 인덱스 (PRIMARY 제외)
-        for idx in table.indexes:
-            if idx.name.upper() != "PRIMARY":
-                col_defs.append(f"    {idx.to_sql_definition(table.name)}")
-
-        # FK
-        for fk in table.foreign_keys:
-            col_defs.append(f"    {fk.to_sql_definition()}")
-
-        lines.append(",\n".join(col_defs))
-        lines.append(f") ENGINE={table.engine} DEFAULT CHARSET={table.charset};")
-
-        return "\n".join(lines)
+from src.core.schema_diff_models import (
+    DiffSeverity,
+    CompareLevel,
+    VersionContext,
+    SeveritySummary,
+    DiffType,
+    ColumnInfo,
+    IndexInfo,
+    ForeignKeyInfo,
+    TableSchema,
+    ColumnDiff,
+    IndexDiff,
+    ForeignKeyDiff,
+    TableDiff,
+    _normalize_column_extra,
+)
+from src.core.schema_extractor import SchemaExtractor
+from src.core.schema_comparator import SchemaComparator
+from src.core.schema_severity_classifier import SeverityClassifier
+from src.core.schema_sync_script_generator import SyncScriptGenerator
